Defining initial pattern, and refactoring main transformation code
