+++ module/compiler/index.ts
@@ -1,8 +1,11 @@
-export * from './src/transform-util';
 export * from './src/util';
 export * from './src/proxy';
 export * from './src/presence';
 export * from './src/module';
-export * from './src/transformers';
 export * from './src/source';
 export * from './src/compiler';
+export * from './src/transformer/transformers';
+export * from './src/transformer/types';
+export * from './src/transformer/state';
+export * from './src/transformer/visitor';
+export * from './src/transformer/transform-util';
