+++ module/di/config/registry.yml
@@ -1,3 +1,6 @@
 ---
 injectable:
    '@travetto/config': Config  
+   '@travetto/di': 
+      - Injectable
+      - Application
