+++ module/test/config/registry.yml
@@ -1,0 +1,6 @@
+--
+test:
+  '@travetto/test': Test
+
+suite:
+  '@travetto/test': Suite
